backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pyjwt
bcrypt
pyyaml
